Moved exclusive BERT methods from HappyTransformer to HappyBERT
